requirements.txt
# ── Core deep learning ────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── MRI file I/O ──────────────────────────────────────────────
nibabel>=5.0.0

# ── Tensor utilities ──────────────────────────────────────────
einops>=0.7.0
numpy>=1.24.0
scipy>=1.10.0

# ── Image metrics ─────────────────────────────────────────────
scikit-image>=0.21.0

# ── Config ────────────────────────────────────────────────────
PyYAML>=6.0

# ── Visualization ─────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Training utilities ────────────────────────────────────────
tqdm>=4.65.0
tensorboard>=2.13.0

# ── Optional: faster data loading ─────────────────────────────
# opencv-python>=4.8.0